Move CVLayer to noise.cv
Also move splitter_symp and invert_permutation to cv.ops
