+++ doc/tutorials/run_error_correction.py
@@ -30,7 +30,6 @@ A complete round of error correction
 import matplotlib.pyplot as plt
 
 from flamingpy.codes import alternating_polarity, SurfaceCode
-from flamingpy.cv.ops import CVLayer
 from flamingpy.decoders import decoder as dec
 from flamingpy.noise import IidNoise
 from flamingpy.utils import viz
