+++ flamingpy/benchmarks/matching.py
@@ -17,10 +17,10 @@ import time
 
 import matplotlib.pyplot as plt
 
-from flamingpy.cv.ops import CVLayer
 from flamingpy.codes import alternating_polarity, SurfaceCode
 from flamingpy.decoders import decoder as dec
 from flamingpy.decoders.mwpm.algos import build_match_graph
+from flamingpy.noise import CVLayer
 
 # How many simulations to do for each algorithm
 num_trials = 10
