+++ flamingpy/noise/__init__.py
@@ -13,3 +13,4 @@
 # limitations under the License.
 """The noise module."""
 from .iid import IidNoise
+from .cv import CVLayer
